Bye bye Typescript
Typescript massively complicated the build chain, and reasons I
originally liked it, such as ES6 modules and single-file compilation,
were quickly taken over by rollup. Futher, the added type safety just
isn't this big a deal at this size, and though it's kind'a nice mostly I
just liked the autocomplete. The last major factor is that anything
in the tool system is a pain for users of the repo; rollup.js is about
the most I'm willing to sacrifice.

The build system was complicated because not because of a single
Typescript flaw but because of a chain of events.

1. Having both Typescript and rollup was basically necessary AFAICT, so
you automatically get a second chain in the link.

2. The two don't play nice, so you have to include sourcery.js, so you
automatically get a third link in the chain.

3. The three things start making you want npm to manage them, so you
semi-automatically get a fourth link in the chain (which also tries to
enter your build system).

4. Then you want to start making Web Workers, which require multiple
files. This is fine with rollup.js, whose config just lets you splat
entry points on. However, Typescript needs special care since you need a
file per worker and you can't squirrel it away in a parallel directory
in any sane manner. Then sourcery.js wants to get involved, and suddenly
you have a redesigned repo layout just to work with Typescript managed
from a shell script or Makefile.

In contrast, with plain JS I can just run rollup --config with a single
config file, and it pretty much just works. The code is shorter, and
the only thing I really miss is the fancier autocomplete.